requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Sockets==0.2.1
gevent==20.6.1
gevent-websocket==0.10.1
greenlet==0.4.16
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
VideoConverter==1.1.0
Werkzeug==1.0.1
zope.event==4.4
zope.interface==5.1.0
